feat(projects): add tech stack verbiage and layout
Adds the tech stack verbiage and list to the
project card.

Also, fixes the layout issue for each project
card where they were not taking the full width
of the parent. This was a bug introduced to fix
the grid layout, but is now resolved using the
default behavior from MaterialUI.