build: update deps (#35)
* build(husky): add execution permissions to hooks

* build: create bun.lockb file

* build(deps): upgrade all deps to latest

* refactor: convert file names to kebab-case

This change converts the file names to use
the kebab-case instead of inconsistently
naming each file based on contents. In
addition to the file name changes, we also
converted the base path to match the new
repository name (i.e., "info").

* ci(workflows): update actions and add cache

This change updates the third-party actions utilized in the ci/cd build. It also introduces a cache
based on the latest information.

refs: https://github.com/actions/starter-workflows/blob/main/pages/nextjs.yml

* fix(husky): remove deprecated hash-bang statements