Feat/shadcn (#39)
* feat(assets): update profile pictures

* feat: add shadcn with tailwind v4

* feat: add initial shadcn layout

* feat(components): add shadcn tooltip

* feat(features): add tooltip for app header social links

* feat(pages): add home link to breadcrumbs

* feat(theme): add theme provider and toggle

This change adds the theme toggle to the app sidebar. This will allow the user to maintain their
existing preferences for their theme when visiting the site.

* fix: add padding div to page layouts

* feat: add react flow to coding languages route

* feat(about): add coding languages graph

* feat(components): add shadcn badge

* fix(app-header): hide social links on mobile

The social links are taking too much header space, so we are hiding when it is mobile.

* fix(app-sidebar): remove unused dependency

* fix(projects): add screen responsive grid for project cards

* feat(projects): add ProjectDetails component

This component will be used to render the formatted project details for each of the projects. This
is the component that is shown when the user selecs See More from the projects card list.

* feat: add initial title for home landing page

* fix: add content to landing/about pages