Feat/imageography screenshots (#41)
* feat(components): add carousel shadcn component

* refactor(assets): remove unused minified-js image

* feat(assets): add imageography screenshots

* feat(features): add github link to project details

* fix(styles): set body color to sidebar background color

This change sets the body color to the background of the sidebar to maintain a consistent background
when overscrolling the main content

* feat(pages): add imageography screenshot carousel

* fix(features): open project github in new tab via anchor target

* feat(features): add project detail with children to projects

This change adds the project detail component that can render a hierarchy of headline/body text
pairings. This change also adds the carousel as the child of the first headline/body pair, so that
we can get the carousel between the overview and the learnings for the imageography section.